+++ src/events.js
@@ -13,10 +13,11 @@ export const CELLS_SELECTION = `${CELLS}.selection`;
 export const CELLS_COLOR = `${CELLS}.color`;
 export const CELLS_HOVER = `${CELLS}.hover`;
 
-// Modifications to the cell set tree structure.
-export const CELL_SETS_MODIFY = `${CELLS}.sets.modify`;
+export const CELL_SETS = `${CELLS}.sets`;
+export const CELL_SETS_ADD = CELL_SETS + ADD;
+export const CELL_SETS_CHANGE = CELL_SETS + CHANGE;
 // Viewing a previously-defined set or group of sets.
-export const CELL_SETS_VIEW = `${CELLS}.sets.view`;
+export const CELL_SETS_VIEW = `${CELL_SETS}.view`;
 
 export const CLUSTERS = 'clusters';
 export const CLUSTERS_ADD = CLUSTERS + ADD;